// src/shared/types/generated-client/models/Host.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 같이 달램 API
 * 모임 매칭 플랫폼 API  ## 인증 방식 - **JWT Bearer Token**: 로그인 후 발급받은 accessToken을 Authorization 헤더에 포함 - accessToken 만료 시간: 15분 - refreshToken 만료 시간: 7일  ## 사용 예시 ``` Authorization: Bearer eyJhbGciOiJIUzI1NiIs... ```
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface Host
 */
export interface Host {
  /**
   *
   * @type {number}
   * @memberof Host
   */
  id: number;
  /**
   *
   * @type {string}
   * @memberof Host
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof Host
   */
  image: string;
}

/**
 * Check if a given object implements the Host interface.
 */
export function instanceOfHost(value: object): value is Host {
  if (!('id' in value) || value['id'] === undefined) return false;
  if (!('name' in value) || value['name'] === undefined) return false;
  if (!('image' in value) || value['image'] === undefined) return false;
  return true;
}

export function HostFromJSON(json: any): Host {
  return HostFromJSONTyped(json, false);
}

export function HostFromJSONTyped(json: any, ignoreDiscriminator: boolean): Host {
  if (json == null) {
    return json;
  }
  return {
    id: json['id'],
    name: json['name'],
    image: json['image'],
  };
}

export function HostToJSON(json: any): Host {
  return HostToJSONTyped(json, false);
}

export function HostToJSONTyped(value?: Host | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    id: value['id'],
    name: value['name'],
    image: value['image'],
  };
}