// src/shared/types/generated-client/models/MeetingList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 같이 달램 API
 * 모임 매칭 플랫폼 API  ## 인증 방식 - **JWT Bearer Token**: 로그인 후 발급받은 accessToken을 Authorization 헤더에 포함 - accessToken 만료 시간: 15분 - refreshToken 만료 시간: 7일  ## 사용 예시 ``` Authorization: Bearer eyJhbGciOiJIUzI1NiIs... ```
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MeetingWithHost } from './MeetingWithHost';
import {
  MeetingWithHostFromJSON,
  MeetingWithHostFromJSONTyped,
  MeetingWithHostToJSON,
  MeetingWithHostToJSONTyped,
} from './MeetingWithHost';

/**
 *
 * @export
 * @interface MeetingList
 */
export interface MeetingList {
  /**
   *
   * @type {Array<MeetingWithHost>}
   * @memberof MeetingList
   */
  data: Array<MeetingWithHost>;
  /**
   *
   * @type {string}
   * @memberof MeetingList
   */
  nextCursor: string;
  /**
   *
   * @type {boolean}
   * @memberof MeetingList
   */
  hasMore: boolean;
}

/**
 * Check if a given object implements the MeetingList interface.
 */
export function instanceOfMeetingList(value: object): value is MeetingList {
  if (!('data' in value) || value['data'] === undefined) return false;
  if (!('nextCursor' in value) || value['nextCursor'] === undefined) return false;
  if (!('hasMore' in value) || value['hasMore'] === undefined) return false;
  return true;
}

export function MeetingListFromJSON(json: any): MeetingList {
  return MeetingListFromJSONTyped(json, false);
}

export function MeetingListFromJSONTyped(json: any, ignoreDiscriminator: boolean): MeetingList {
  if (json == null) {
    return json;
  }
  return {
    data: (json['data'] as Array<any>).map(MeetingWithHostFromJSON),
    nextCursor: json['nextCursor'],
    hasMore: json['hasMore'],
  };
}

export function MeetingListToJSON(json: any): MeetingList {
  return MeetingListToJSONTyped(json, false);
}

export function MeetingListToJSONTyped(
  value?: MeetingList | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    data: (value['data'] as Array<any>).map(MeetingWithHostToJSON),
    nextCursor: value['nextCursor'],
    hasMore: value['hasMore'],
  };
}